refactor(synth): move signal out of util
