+++ packages/synth/src/automation/event-timeline.ts
@@ -1,7 +1,7 @@
 import { assertDefined, assertUnreachable, assertedAt } from '@repo/common/assert'
 import { isNonEmpty } from '@repo/common/predicate'
+import { Signal } from '#signal'
 import type { SynthTime } from '#time'
-import { Signal } from '#util/signal'
 import type { ReadonlyEventTimeline, TimedEvent } from './readonly-event-timeline'
 
 export class EventTimeline<TEvent extends TimedEvent> implements ReadonlyEventTimeline<TEvent> {
