+++ packages/synth/src/signal/index.ts
