+++ packages/synth/src/signal/signal.ts
